requirements.txt
# Essential LangChain components
langchain
langchain-core
langchain-google-genai
langchain-google-vertexai
langgraph
google-generativeai
google-cloud-aiplatform

# Web UI
streamlit
python-dotenv

# Basic utilities
requests
aiohttp
pydantic

# PDF processing
PyPDF2

# Vector stores and embeddings for RAG
faiss-cpu
sentence-transformers
chromadb
tiktoken

# Document processing
pypdf
python-docx
pandas
openpyxl

# Optional: Development tools
# pytest
# black 